# Clean up asm