--- a/murmur32_amd64.s
+++ b/murmur32_amd64.s
@@ -1,100 +1,95 @@
 // +build go1.5,amd64
 
+// Sum32(data []byte) (h1 uint32)
+TEXT ·Sum32(SB), $0-28
+	MOVQ data_base+0(FP), SI
+	MOVQ data_len+8(FP), R9
+	XORL R12, R12
+	LEAQ h1+24(FP), BX
+	JMP  sum32internal<>(SB)
+
+// SeedSum32(seed uint32, data []byte) (h1 uint32)
+TEXT ·SeedSum32(SB), $0-36
+	MOVL seed+0(FP), R12
+	MOVQ data_base+8(FP), SI
+	MOVQ data_len+16(FP), R9
+	LEAQ h1+32(FP), BX
+	JMP  sum32internal<>(SB)
+
 #define c1_32 0xcc9e2d51
 #define c2_32 0x1b873593
 
-// AX = h1
-// SI = &data[0]
+// Expects:
+// R12 == uint32 seed
+// SI  == &data
+// R9  == len(data)
+// BX  == &uint32 return
+TEXT sum32internal<>(SB), $0
+	XORQ BP, BP
+	MOVQ R9, CX
+	ANDQ $-4, CX // cx = len(data) - (len(data) % 4)
 
-#define sum32loop \
-	MOVQ  CX, DX;                   \
-	SHRQ  $2, DX;                   \
-	SHLQ  $2, DX;                   \ // DX = end
-	MOVQ  $0, BP;                   \ // BP = word pointer
-	loop:                           \
-	CMPQ  BP, DX;                   \
-	JE    end                       \
-	                                \
-	MOVL  (SI)(BP*1), DI;           \ // DI = k1
-	ADDQ  $4, BP;                   \
-	IMULL $c1_32, DI;               \
-	ROLL  $15, DI;                  \
-	IMULL $c2_32, DI;               \
-	XORL  AX, DI;                   \
-	ROLL  $13, DI;                  \
-	LEAL  0xe6546b64(DI)(DI*4), AX; \
-	JMP   loop;                     \
-	      \
-	end:  \
+	// for bp = 0 ; bp < cx; bp += 4 {...
+loop:
+	CMPQ BP, CX
+	JE   tail
+	MOVL (SI)(BP*1), AX
+	ADDQ $4, BP
 
-// DX/DI = scratch
-// BX = original clen
-#define sum32tail \
-	MOVL    $0, DX;    \ // DX is k1, DI is intermediate calc
-	CMPQ    CX, $3;    \
-	JNE     chk2;      \
-	                   \
-	on3:               \
-	MOVBLZX 2(SI), DI; \
-	SHLL    $16, DI;   \
-	XORL    DI, DX;    \
-	JMP     on2;       \
-	                   \
-	chk2:              \
-	CMPQ    CX, $2;    \
-	JNE     chk1;      \
-	                   \
-	on2:               \
-	MOVBLZX 1(SI), DI; \
-	SHLL    $8, DI;    \
-	XORL    DI, DX;    \
-	JMP     on1;       \
-	                         \
-	chk1:                    \
-	CMPQ    CX, $1;          \
-	JNE     chk0;            \
-	                         \
-	on1:                     \
-	MOVBLZX (SI), DI;        \
-	XORL    DI, DX;          \
-	IMULL   $c1_32, DX;      \
-	ROLL    $15, DX;         \
-	IMULL   $c2_32, DX;      \
-	XORL    DX, AX;          \
-	                         \
-	chk0:                    \
-	XORL    BX, AX;          \
-	MOVL    AX, DI;          \
-	SHRL    $16, DI;         \
-	XORL    DI, AX;          \
-	IMULL   $0x85ebca6b, AX; \
-	MOVL    AX, DI;          \
-	SHRL    $13, DI;         \
-	XORL    DI, AX;          \
-	IMULL   $0xc2b2ae35, AX; \
-	MOVL    AX, DI;          \
-	SHRL    $16, DI;         \
-	XORL    DI, AX;          \
+	IMULL $c1_32, AX
+	ROLL  $15, AX
+	IMULL $c2_32, AX
 
-#define sum32 \
-	sum32loop    \
-	MOVQ CX, BX; \ // save clen for sum32tail
-	ADDQ DX, SI; \ // adjust SI to partial block
-	SUBQ DX, CX; \ // adjust CX to partial block len
-	sum32tail    \
+	XORL AX, R12
+	ROLL $13, R12
+	LEAL 0xe6546b64(R12)(R12*4), R12
 
-TEXT ·Sum32(SB), $0-28
-	MOVL $0, AX
-	MOVQ data_base+0(FP), SI
-	MOVQ data_len+8(FP), CX
-	sum32
-	MOVL AX, h1+24(FP)
-	RET
+	JNE loop
 
-TEXT ·SeedSum32(SB), $0-36
-	MOVL seed+0(FP), AX
-	MOVQ data_base+8(FP), SI
-	MOVQ data_len+16(FP), CX
-	sum32
-	MOVL AX, h1+32(FP)
+tail:
+	MOVQ R9, CX
+	ANDL $3, CX
+	JZ   finalize
+
+	XORQ AX, AX
+
+	DECL CX
+	JZ   tail1
+	DECL CX
+	JZ   tail2
+
+tail3:  // no jump
+	MOVB 2(SI)(BP*1), AX
+	SALL $16, AX
+
+tail2:
+	MOVW (SI)(BP*1), AX
+	JMP  fintail
+
+tail1:
+	MOVB (SI)(BP*1), AL
+
+fintail:
+	IMULL $c1_32, AX
+	ROLL  $15, AX
+	IMULL $c2_32, AX
+	XORL  AX, R12
+
+finalize:
+	XORL R9, R12
+
+	// fmix32
+	MOVL  R12, DX
+	SHRL  $16, DX
+	XORL  DX, R12
+	IMULL $0x85ebca6b, R12
+	MOVL  R12, DX
+	SHRL  $13, DX
+	XORL  DX, R12
+	IMULL $0xc2b2ae35, R12
+	MOVL  R12, DX
+	SHRL  $16, DX
+	XORL  DX, R12
+
+	MOVL R12, (BX)
 	RET