on an Intel(R) Xeon(R) CPU E5-2690 v2 @ 3.00GHz
```
on an Intel(R) Xeon(R) CPU E5-2690 v2 @ 3.00GHz
benchmark                       old ns/op     new ns/op     delta
BenchmarkBase32-40              6.05          5.71          -5.62%
BenchmarkLoop32-40              6.13          5.75          -6.20%
BenchmarkTail32_1-40            6.39          6.42          +0.47%
BenchmarkTail32_2-40            6.21          5.75          -7.41%
BenchmarkTail32_3-40            6.49          5.79          -10.79%
BenchmarkBasePartial32-40       113           116           +2.65%
BenchmarkLoopPartial32-40       114           119           +4.39%
BenchmarkTailPartial32_1-40     114           118           +3.51%
BenchmarkTailPartial32_2-40     117           120           +2.56%
BenchmarkTailPartial32_3-40     119           121           +1.68%
BenchmarkBase128-40             11.8          7.40          -37.29%
BenchmarkLoop128-40             13.2          9.91          -24.92%
BenchmarkTail128_1-40           11.8          8.42          -28.64%
BenchmarkTail128_2-40           11.8          8.80          -25.42%
BenchmarkTail128_3-40           11.8          9.14          -22.54%
BenchmarkTail128_4-40           11.5          9.51          -17.30%
BenchmarkTail128_5-40           11.8          10.2          -13.56%
BenchmarkTail128_6-40           11.5          10.2          -11.30%
BenchmarkTail128_7-40           11.5          10.8          -6.09%
BenchmarkTail128_8-40           11.7          10.5          -10.26%
BenchmarkTail128_9-40           12.6          11.2          -11.11%
BenchmarkTail128_10-40          12.9          11.8          -8.53%
BenchmarkTail128_11-40          12.9          12.2          -5.43%
BenchmarkTail128_12-40          13.0          12.2          -6.15%
BenchmarkTail128_13-40          13.3          13.1          -1.50%
BenchmarkTail128_14-40          13.5          13.2          -2.22%
BenchmarkTail128_15-40          13.6          13.2          -2.94%
```