@jeddenlea 2016-05-19
```
[jed@lappy]tmp$ benchcmp {old,new}_xeon
benchmark                old ns/op     new ns/op     delta
Benchmark32_1-40         5.83          6.40          +9.78%
Benchmark32_2-40         6.23          5.73          -8.03%
Benchmark32_4-40         6.15          5.39          -12.36%
Benchmark32_8-40         7.76          6.78          -12.63%
Benchmark32_16-40        10.5          9.28          -11.62%
Benchmark32_32-40        15.6          14.6          -6.41%
Benchmark32_64-40        27.7          26.7          -3.61%
Benchmark32_128-40       50.5          50.2          -0.59%
Benchmark32_256-40       110           107           -2.73%
Benchmark32_512-40       217           216           -0.46%
Benchmark32_1024-40      433           432           -0.23%
Benchmark32_2048-40      865           864           -0.12%
Benchmark32_4096-40      1727          1728          +0.06%
Benchmark32_8192-40      3453          3457          +0.12%
Benchmark128_1-40        11.8          8.44          -28.47%
Benchmark128_2-40        11.5          8.78          -23.65%
Benchmark128_4-40        11.8          9.46          -19.83%
Benchmark128_8-40        11.6          10.5          -9.48%
Benchmark128_16-40       13.5          9.95          -26.30%
Benchmark128_32-40       15.2          12.7          -16.45%
Benchmark128_64-40       21.0          19.2          -8.57%
Benchmark128_128-40      31.3          30.0          -4.15%
Benchmark128_256-40      53.6          52.8          -1.49%
Benchmark128_512-40      104           99.0          -4.81%
Benchmark128_1024-40     195           190           -2.56%
Benchmark128_2048-40     376           371           -1.33%
Benchmark128_4096-40     739           734           -0.68%
Benchmark128_8192-40     1463          1461          -0.14%

benchmark                old MB/s     new MB/s     speedup
Benchmark32_1-40         171.51       156.16       0.91x
Benchmark32_2-40         321.08       348.87       1.09x
Benchmark32_4-40         650.59       742.48       1.14x
Benchmark32_8-40         1030.63      1179.33      1.14x
Benchmark32_16-40        1524.10      1723.38      1.13x
Benchmark32_32-40        2049.80      2185.86      1.07x
Benchmark32_64-40        2313.08      2395.27      1.04x
Benchmark32_128-40       2533.36      2548.71      1.01x
Benchmark32_256-40       2323.06      2371.06      1.02x
Benchmark32_512-40       2349.99      2369.30      1.01x
Benchmark32_1024-40      2361.12      2369.38      1.00x
Benchmark32_2048-40      2367.23      2369.42      1.00x
Benchmark32_4096-40      2370.77      2369.14      1.00x
Benchmark32_8192-40      2372.26      2369.31      1.00x
Benchmark128_1-40        84.75        118.45       1.40x
Benchmark128_2-40        174.46       227.81       1.31x
Benchmark128_4-40        339.19       423.01       1.25x
Benchmark128_8-40        687.52       764.51       1.11x
Benchmark128_16-40       1186.63      1607.73      1.35x
Benchmark128_32-40       2109.00      2522.91      1.20x
Benchmark128_64-40       3043.83      3330.73      1.09x
Benchmark128_128-40      4083.94      4268.56      1.05x
Benchmark128_256-40      4774.09      4849.73      1.02x
Benchmark128_512-40      4901.69      5170.41      1.05x
Benchmark128_1024-40     5241.77      5379.17      1.03x
Benchmark128_2048-40     5439.65      5506.11      1.01x
Benchmark128_4096-40     5540.57      5574.32      1.01x
Benchmark128_8192-40     5596.12      5604.63      1.00x
```