@twmb 2016-05-20
On n1-standard-2 us-central1-b compute instance, 1.6.2:
```
benchmark               old ns/op     new ns/op     delta
Benchmark32_1-2         6.18          5.38          -12.94%
Benchmark32_2-2         6.47          5.44          -15.92%
Benchmark32_4-2         6.38          5.83          -8.62%
Benchmark32_8-2         7.68          7.41          -3.52%
Benchmark32_16-2        10.6          9.65          -8.96%
Benchmark32_22-2        12.3          12.0          -2.44%
Benchmark32_32-2        15.6          15.6          +0.00%
Benchmark32_64-2        28.8          28.8          +0.00%
Benchmark32_128-2       52.7          53.2          +0.95%
Benchmark32_256-2       117           117           +0.00%
Benchmark32_512-2       233           236           +1.29%
Benchmark32_1024-2      468           476           +1.71%
Benchmark32_2048-2      937           953           +1.71%
Benchmark32_4096-2      1873          1903          +1.60%
Benchmark32_8192-2      3757          3822          +1.73%
Benchmark32_16k-2       7521          7611          +1.20%
Benchmark32_32k-2       14980         15228         +1.66%
Benchmark128_1-2        10.4          8.93          -14.13%
Benchmark128_2-2        10.8          9.34          -13.52%
Benchmark128_4-2        11.1          9.71          -12.52%
Benchmark128_8-2        12.2          10.0          -18.03%
Benchmark128_16-2       11.8          10.2          -13.56%
Benchmark128_22-2       13.3          12.2          -8.27%
Benchmark128_32-2       14.0          13.0          -7.14%
Benchmark128_64-2       21.3          18.7          -12.21%
Benchmark128_128-2      33.3          31.5          -5.41%
Benchmark128_256-2      55.5          54.9          -1.08%
Benchmark128_512-2      102           104           +1.96%
Benchmark128_1024-2     201           199           -1.00%
Benchmark128_2048-2     391           389           -0.51%
Benchmark128_4096-2     776           769           -0.90%
Benchmark128_8192-2     1529          1530          +0.07%
Benchmark128_16k-2      3055          3057          +0.07%
Benchmark128_32k-2      6123          6149          +0.42%

benchmark               old MB/s     new MB/s     speedup
Benchmark32_1-2         161.77       185.80       1.15x
Benchmark32_2-2         309.12       367.38       1.19x
Benchmark32_4-2         626.92       685.68       1.09x
Benchmark32_8-2         1041.08      1079.62      1.04x
Benchmark32_16-2        1502.35      1658.83      1.10x
Benchmark32_22-2        1788.02      1829.25      1.02x
Benchmark32_32-2        2044.89      2054.11      1.00x
Benchmark32_64-2        2218.88      2224.12      1.00x
Benchmark32_128-2       2431.06      2404.23      0.99x
Benchmark32_256-2       2183.15      2181.80      1.00x
Benchmark32_512-2       2191.57      2162.69      0.99x
Benchmark32_1024-2      2187.56      2148.05      0.98x
Benchmark32_2048-2      2183.51      2148.76      0.98x
Benchmark32_4096-2      2186.01      2152.00      0.98x
Benchmark32_8192-2      2180.32      2143.16      0.98x
Benchmark32_16k-2       2178.18      2152.66      0.99x
Benchmark32_32k-2       2187.45      2151.80      0.98x
Benchmark128_1-2        95.91        111.99       1.17x
Benchmark128_2-2        184.70       214.09       1.16x
Benchmark128_4-2        361.29       412.01       1.14x
Benchmark128_8-2        653.28       797.27       1.22x
Benchmark128_16-2       1354.09      1573.96      1.16x
Benchmark128_22-2       1657.03      1804.62      1.09x
Benchmark128_32-2       2291.89      2456.17      1.07x
Benchmark128_64-2       3002.45      3419.30      1.14x
Benchmark128_128-2      3848.38      4069.82      1.06x
Benchmark128_256-2      4615.59      4664.45      1.01x
Benchmark128_512-2      4979.39      4882.39      0.98x
Benchmark128_1024-2     5085.26      5121.31      1.01x
Benchmark128_2048-2     5228.19      5256.75      1.01x
Benchmark128_4096-2     5273.05      5321.84      1.01x
Benchmark128_8192-2     5356.31      5353.50      1.00x
Benchmark128_16k-2      5361.54      5358.48      1.00x
Benchmark128_32k-2      5350.79      5328.86      1.00x
```

same instance, tip: (`go version devel +85e39f8 Fri May 20 17:10:18 2016 +0000 linux/amd64`):
```
benchmark               old ns/op     new ns/op     delta
Benchmark32_1-2         5.97          5.30          -11.22%
Benchmark32_2-2         6.28          5.45          -13.22%
Benchmark32_4-2         6.15          5.58          -9.27%
Benchmark32_8-2         7.61          6.80          -10.64%
Benchmark32_16-2        10.3          9.63          -6.50%
Benchmark32_22-2        12.1          11.9          -1.65%
Benchmark32_32-2        15.7          15.2          -3.18%
Benchmark32_64-2        28.8          29.1          +1.04%
Benchmark32_128-2       59.3          52.7          -11.13%
Benchmark32_256-2       117           115           -1.71%
Benchmark32_512-2       235           236           +0.43%
Benchmark32_1024-2      470           470           +0.00%
Benchmark32_2048-2      933           944           +1.18%
Benchmark32_4096-2      1877          1875          -0.11%
Benchmark32_8192-2      3750          3781          +0.83%
Benchmark32_16k-2       7590          7603          +0.17%
Benchmark32_32k-2       15188         15188         +0.00%
Benchmark128_1-2        10.4          9.30          -10.58%
Benchmark128_2-2        10.7          9.65          -9.81%
Benchmark128_4-2        11.1          10.1          -9.01%
Benchmark128_8-2        12.2          10.4          -14.75%
Benchmark128_16-2       12.3          10.0          -18.70%
Benchmark128_22-2       13.5          12.3          -8.89%
Benchmark128_32-2       14.2          12.7          -10.56%
Benchmark128_64-2       21.0          18.4          -12.38%
Benchmark128_128-2      32.9          31.2          -5.17%
Benchmark128_256-2      55.1          54.5          -1.09%
Benchmark128_512-2      103           102           -0.97%
Benchmark128_1024-2     202           198           -1.98%
Benchmark128_2048-2     391           391           +0.00%
Benchmark128_4096-2     776           765           -1.42%
Benchmark128_8192-2     1544          1527          -1.10%
Benchmark128_16k-2      3083          3042          -1.33%
Benchmark128_32k-2      6068          6097          +0.48%

benchmark               old MB/s     new MB/s     speedup
Benchmark32_1-2         167.49       188.56       1.13x
Benchmark32_2-2         318.49       367.26       1.15x
Benchmark32_4-2         650.48       716.27       1.10x
Benchmark32_8-2         1051.20      1175.70      1.12x
Benchmark32_16-2        1556.90      1661.55      1.07x
Benchmark32_22-2        1814.90      1850.16      1.02x
Benchmark32_32-2        2043.36      2101.63      1.03x
Benchmark32_64-2        2222.83      2198.38      0.99x
Benchmark32_128-2       2159.35      2426.97      1.12x
Benchmark32_256-2       2177.89      2220.65      1.02x
Benchmark32_512-2       2171.84      2166.90      1.00x
Benchmark32_1024-2      2175.37      2175.00      1.00x
Benchmark32_2048-2      2194.89      2168.47      0.99x
Benchmark32_4096-2      2181.72      2183.53      1.00x
Benchmark32_8192-2      2184.08      2166.15      0.99x
Benchmark32_16k-2       2158.53      2154.76      1.00x
Benchmark32_32k-2       2157.38      2157.45      1.00x
Benchmark128_1-2        96.46        107.58       1.12x
Benchmark128_2-2        186.79       207.27       1.11x
Benchmark128_4-2        360.70       397.63       1.10x
Benchmark128_8-2        657.37       766.56       1.17x
Benchmark128_16-2       1301.72      1594.44      1.22x
Benchmark128_22-2       1633.83      1794.39      1.10x
Benchmark128_32-2       2250.64      2514.41      1.12x
Benchmark128_64-2       3048.43      3469.59      1.14x
Benchmark128_128-2      3885.10      4097.30      1.05x
Benchmark128_256-2      4643.92      4694.06      1.01x
Benchmark128_512-2      4962.42      5017.08      1.01x
Benchmark128_1024-2     5053.39      5153.70      1.02x
Benchmark128_2048-2     5228.93      5235.24      1.00x
Benchmark128_4096-2     5277.55      5350.58      1.01x
Benchmark128_8192-2     5302.86      5361.35      1.01x
Benchmark128_16k-2      5313.45      5385.51      1.01x
Benchmark128_32k-2      5399.74      5374.36      1.00x
```

Intel 32 core @ CPU2899.660MHz, tip still:
```
benchmark                old ns/op     new ns/op     delta
Benchmark32_1-32         6.97          6.87          -1.43%
Benchmark32_2-32         7.47          6.30          -15.66%
Benchmark32_4-32         7.74          6.67          -13.82%
Benchmark32_8-32         9.38          8.43          -10.13%
Benchmark32_16-32        13.2          11.8          -10.61%
Benchmark32_22-32        14.8          14.4          -2.70%
Benchmark32_32-32        19.0          19.0          +0.00%
Benchmark32_64-32        33.7          33.1          -1.78%
Benchmark32_128-32       65.9          61.9          -6.07%
Benchmark32_256-32       125           121           -3.20%
Benchmark32_512-32       245           238           -2.86%
Benchmark32_1024-32      478           475           -0.63%
Benchmark32_2048-32      938           945           +0.75%
Benchmark32_4096-32      1892          1921          +1.53%
Benchmark32_8192-32      3742          3822          +2.14%
Benchmark32_16k-32       7451          7579          +1.72%
Benchmark32_32k-32       14981         15189         +1.39%
Benchmark128_1-32        12.9          10.4          -19.38%
Benchmark128_2-32        13.3          10.5          -21.05%
Benchmark128_4-32        13.5          11.1          -17.78%
Benchmark128_8-32        13.8          11.8          -14.49%
Benchmark128_16-32       14.9          12.0          -19.46%
Benchmark128_22-32       16.4          14.6          -10.98%
Benchmark128_32-32       17.8          15.4          -13.48%
Benchmark128_64-32       23.3          21.9          -6.01%
Benchmark128_128-32      35.9          35.4          -1.39%
Benchmark128_256-32      62.0          61.7          -0.48%
Benchmark128_512-32      114           115           +0.88%
Benchmark128_1024-32     223           219           -1.79%
Benchmark128_2048-32     434           428           -1.38%
Benchmark128_4096-32     849           843           -0.71%
Benchmark128_8192-32     1688          1706          +1.07%
Benchmark128_16k-32      3360          3401          +1.22%
Benchmark128_32k-32      6690          6794          +1.55%

benchmark                old MB/s     new MB/s     speedup
Benchmark32_1-32         143.41       145.54       1.01x
Benchmark32_2-32         267.80       317.46       1.19x
Benchmark32_4-32         517.03       599.57       1.16x
Benchmark32_8-32         852.84       948.92       1.11x
Benchmark32_16-32        1214.03      1356.52      1.12x
Benchmark32_22-32        1483.75      1530.79      1.03x
Benchmark32_32-32        1680.35      1681.65      1.00x
Benchmark32_64-32        1897.88      1933.61      1.02x
Benchmark32_128-32       1942.14      2068.34      1.06x
Benchmark32_256-32       2035.65      2105.69      1.03x
Benchmark32_512-32       2088.90      2150.75      1.03x
Benchmark32_1024-32      2139.77      2153.19      1.01x
Benchmark32_2048-32      2181.50      2166.72      0.99x
Benchmark32_4096-32      2164.43      2131.64      0.98x
Benchmark32_8192-32      2189.19      2143.28      0.98x
Benchmark32_16k-32       2198.88      2161.70      0.98x
Benchmark32_32k-32       2187.28      2157.31      0.99x
Benchmark128_1-32        77.41        96.49        1.25x
Benchmark128_2-32        150.28       191.26       1.27x
Benchmark128_4-32        296.67       361.54       1.22x
Benchmark128_8-32        578.18       679.81       1.18x
Benchmark128_16-32       1073.54      1338.16      1.25x
Benchmark128_22-32       1342.88      1502.13      1.12x
Benchmark128_32-32       1796.93      2075.40      1.15x
Benchmark128_64-32       2741.23      2916.60      1.06x
Benchmark128_128-32      3566.57      3610.72      1.01x
Benchmark128_256-32      4129.27      4146.72      1.00x
Benchmark128_512-32      4470.79      4446.87      0.99x
Benchmark128_1024-32     4575.14      4661.00      1.02x
Benchmark128_2048-32     4713.71      4781.43      1.01x
Benchmark128_4096-32     4822.27      4854.51      1.01x
Benchmark128_8192-32     4851.12      4800.75      0.99x
Benchmark128_16k-32      4875.35      4817.20      0.99x
Benchmark128_32k-32      4897.98      4822.98      0.98x
```

I'll review this weekend. Just wanted to grab a few more benchmarks specifically against tip.