@twmb 2016-05-21
Running benchmarks against PR #5 on the 32 core server demonstrates the strengths of these tail optimizations even better:

```
benchmark                       old ns/op     new ns/op     delta
BenchmarkBase32-32              6.32          6.06          -4.11%
BenchmarkLoop32-32              7.65          6.71          -12.29%
BenchmarkTail32_1-32            7.09          6.68          -5.78%
BenchmarkTail32_2-32            7.37          6.43          -12.75%
BenchmarkTail32_3-32            8.06          6.69          -17.00%
BenchmarkBasePartial32-32       112           112           +0.00%
BenchmarkLoopPartial32-32       114           114           +0.00%
BenchmarkTailPartial32_1-32     114           114           +0.00%
BenchmarkTailPartial32_2-32     115           115           +0.00%
BenchmarkTailPartial32_3-32     119           118           -0.84%
BenchmarkBase128-32             13.9          8.44          -39.28%
BenchmarkLoop128-32             15.5          11.9          -23.23%
BenchmarkTail128_1-32           14.2          10.0          -29.58%
BenchmarkTail128_2-32           14.2          10.3          -27.46%
BenchmarkTail128_3-32           13.8          10.7          -22.46%
BenchmarkTail128_4-32           14.0          11.2          -20.00%
BenchmarkTail128_5-32           14.0          12.0          -14.29%
BenchmarkTail128_6-32           14.0          11.7          -16.43%
BenchmarkTail128_7-32           13.8          12.3          -10.87%
BenchmarkTail128_8-32           13.8          12.2          -11.59%
BenchmarkTail128_9-32           15.0          13.6          -9.33%
BenchmarkTail128_10-32          15.3          13.6          -11.11%
BenchmarkTail128_11-32          15.5          14.3          -7.74%
BenchmarkTail128_12-32          15.9          14.3          -10.06%
BenchmarkTail123_13-32          16.0          15.0          -6.25%
BenchmarkTail128_14-32          16.2          16.1          -0.62%
BenchmarkTail128_15-32          16.4          15.6          -4.88%

benchmark                       old MB/s     new MB/s     speedup
BenchmarkLoop32-32              522.59       596.43       1.14x
BenchmarkTail32_1-32            141.10       149.62       1.06x
BenchmarkTail32_2-32            271.41       310.86       1.15x
BenchmarkTail32_3-32            372.41       448.25       1.20x
BenchmarkLoopPartial32-32       34.81        34.87        1.00x
BenchmarkTailPartial32_1-32     8.72         8.74         1.00x
BenchmarkTailPartial32_2-32     17.38        17.33        1.00x
BenchmarkTailPartial32_3-32     25.13        25.31        1.01x
BenchmarkLoop128-32             1032.35      1348.84      1.31x
BenchmarkTail128_1-32           70.59        99.62        1.41x
BenchmarkTail128_2-32           140.61       193.89       1.38x
BenchmarkTail128_3-32           217.17       281.36       1.30x
BenchmarkTail128_4-32           285.59       358.74       1.26x
BenchmarkTail128_5-32           357.00       418.05       1.17x
BenchmarkTail128_6-32           428.72       513.69       1.20x
BenchmarkTail128_7-32           505.43       569.13       1.13x
BenchmarkTail128_8-32           578.98       656.32       1.13x
BenchmarkTail128_9-32           598.69       661.79       1.11x
BenchmarkTail128_10-32          651.76       734.21       1.13x
BenchmarkTail128_11-32          707.49       770.08       1.09x
BenchmarkTail128_12-32          754.05       839.56       1.11x
BenchmarkTail123_13-32          814.54       865.33       1.06x
BenchmarkTail128_14-32          865.42       869.05       1.00x
BenchmarkTail128_15-32          914.61       962.90       1.05x
```